migrate vtclient to cobra and rewire tests
