+++ go/cmd/vtclient/cli/plugin_opentracing.go
@@ -14,11 +14,10 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package cli
 
 import (
 	"vitess.io/vitess/go/trace"
-
 	"vitess.io/vitess/go/vt/servenv"
 )
